Cleanup and extract strings for localization.
