+++ libs/SalesforceSDK/res/values/sf__strings.xml
@@ -59,7 +59,13 @@
 
 	<!-- ScreenLock -->
 	<string name="sf__logout">Logout</string>
+	<string name="sf__application_icon">Application Icon</string>
+	<string name="sf__screen_lock_auth_failed">Authentication Failed.</string>
+	<string name="sf__screen_lock_auth_error">Authentication error.</string>
+	<string name="sf__screen_lock_auth_success">Authentication error.</string>
+	<string name="sf__screen_lock_title">%s Lock</string>
+	<string name="sf__screen_lock_subtitle">Unlock to access your %s account.</string>
 
 	<!-- Dev support -->
-	<string name="sf__dev_support_title">Mobile SDK Dev Support</string>
+	<string name="sf__dev_support_title">Authentication succeeded.</string>
 </resources>
